+++ obbm-be/src/main/java/com/springboot/obbm/dto/payment/request/CreatePaymentLinkRequest.java
@@ -3,6 +3,8 @@ package com.springboot.obbm.dto.payment.request;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,5 +14,5 @@ public class CreatePaymentLinkRequest {
     Integer contractId;
     String productName;
     String description;
-    Double prepay;
+    Long prepay;
 }
